% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_discretizer.R
\docType{class}
\name{Information_Value_Discretizer-class}
\alias{Information_Value_Discretizer}
\alias{Information_Value_Discretizer-class}
\title{Information Value Discretizer}
\section{Methods}{

\describe{
\item{\code{fit(x, y)}}{Discretize a numeric variable using information value
  \subsection{Parameters}{
    \itemize{
      \item{\code{x} numeric field to discretize.}
      \item{\code{y} factor with 2 levels used to guide discretization.}
    }
  }
  \subsection{Value}{
    Modifies the object in place to generate break intervals.
  }}
}}

