% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_discretizer.R
\docType{class}
\name{Information_Value_Discretizer-class}
\alias{Information_Value_Discretizer}
\alias{Information_Value_Discretizer-class}
\title{Information Value Discretizer}
\section{Methods}{

\describe{
\item{\code{metric_(z)}}{Metric calculated on a summary table of x and y}

\item{\code{metric_list_of_tables_(l)}}{Metric calculated on a list of tables}
}}

